Added `Get_Quiz_Capable_Trait`
Refactored code to use it.
